refactor all the structs
