+++ programs/virtual-curve/src/params/fee_parameters.rs
@@ -5,7 +5,6 @@ use crate::constants::fee::{
 };
 use crate::constants::{BASIS_POINT_MAX, BIN_STEP_BPS_DEFAULT, BIN_STEP_BPS_U128_DEFAULT, U24_MAX};
 use crate::error::PoolError;
-use crate::state::fee::{BaseFeeStruct, DynamicFeeStruct, PoolFeesStruct};
 use crate::state::{BaseFeeConfig, DynamicFeeConfig, PoolFeesConfig};
 use anchor_lang::prelude::*;
 
@@ -29,10 +28,10 @@ pub struct BaseFeeParameters {
 
 impl BaseFeeParameters {
     fn validate(&self) -> Result<()> {
-        let base_fee_struct = self.to_base_fee_struct();
+        let base_fee_config = self.to_base_fee_config();
 
-        let min_fee_numerator = base_fee_struct.get_min_base_fee_numerator()?;
-        let max_fee_numerator = base_fee_struct.get_max_base_fee_numerator();
+        let min_fee_numerator = base_fee_config.get_min_base_fee_numerator()?;
+        let max_fee_numerator = base_fee_config.get_max_base_fee_numerator();
         validate_fee_fraction(min_fee_numerator, FEE_DENOMINATOR)?;
         validate_fee_fraction(max_fee_numerator, FEE_DENOMINATOR)?;
         require!(
@@ -40,17 +39,6 @@ impl BaseFeeParameters {
             PoolError::ExceedMaxFeeBps
         );
         Ok(())
-    }
-
-    fn to_base_fee_struct(&self) -> BaseFeeStruct {
-        BaseFeeStruct {
-            cliff_fee_numerator: self.cliff_fee_numerator,
-            number_of_period: self.number_of_period,
-            period_frequency: self.period_frequency,
-            reduction_factor: self.reduction_factor,
-            fee_scheduler_mode: self.fee_scheduler_mode,
-            ..Default::default()
-        }
     }
 
     pub fn to_base_fee_config(&self) -> BaseFeeConfig {
@@ -88,29 +76,6 @@ impl PoolFeeParamters {
             }
         }
     }
-
-    pub fn to_pool_fees_struct(&self) -> PoolFeesStruct {
-        let &PoolFeeParamters {
-            base_fee,
-            dynamic_fee,
-        } = self;
-        if let Some(dynamic_fee) = dynamic_fee {
-            PoolFeesStruct {
-                base_fee: base_fee.to_base_fee_struct(),
-                protocol_fee_percent: PROTOCOL_FEE_PERCENT,
-                referral_fee_percent: HOST_FEE_PERCENT,
-                dynamic_fee: dynamic_fee.to_dynamic_fee_struct(),
-                ..Default::default()
-            }
-        } else {
-            PoolFeesStruct {
-                base_fee: base_fee.to_base_fee_struct(),
-                protocol_fee_percent: PROTOCOL_FEE_PERCENT,
-                referral_fee_percent: HOST_FEE_PERCENT,
-                ..Default::default()
-            }
-        }
-    }
 }
 
 #[derive(Copy, Clone, Debug, AnchorSerialize, AnchorDeserialize, InitSpace, Default)]
@@ -133,20 +98,6 @@ impl DynamicFeeParameters {
             decay_period: self.decay_period,
             reduction_factor: self.reduction_factor,
             bin_step_u128: self.bin_step_u128,
-            max_volatility_accumulator: self.max_volatility_accumulator,
-            variable_fee_control: self.variable_fee_control,
-            ..Default::default()
-        }
-    }
-
-    fn to_dynamic_fee_struct(&self) -> DynamicFeeStruct {
-        DynamicFeeStruct {
-            initialized: 1,
-            bin_step: self.bin_step,
-            bin_step_u128: self.bin_step_u128,
-            filter_period: self.filter_period,
-            decay_period: self.decay_period,
-            reduction_factor: self.reduction_factor,
             max_volatility_accumulator: self.max_volatility_accumulator,
             variable_fee_control: self.variable_fee_control,
             ..Default::default()
